Pushing From the Weekend
I think we have the updates done.
Tweaked the main hazard function, does a type and safety check on omega
and updates the structure if necessary.
Tweaked male/female mortality to handle new shape (males) and new stages
(females).
Need to continue testing to make sure edge cases are successful.
Started setting up a vignette to test that the changes are successful.