Cubes are Finished
Finalized the documentation - fixed a lot of things so they actually
render properly, which means all of my original work probably doesn't
render right. Oops.
Had to copy a script from V1 to finish setting up cubes. We never
exported it because cubes only existed within V1, and V2 was an
extension of that, so it never had its own cubes.
As far as I can tell, both cubes are ready to go.