Updates and Upgrades
Rebuilt the releases and batch migration structures to be more
efficient.
Updated trajectory documentation to import functions properly.
Created a function to check the markings on the initial petri net.
Re-implemented releases so we can extend them for more functionality.
This is a purely internal change, completely backwards compatible. This
extension allows us to perform small-molecule style releases. It should
also make it easier to extend releases in the future.
Ran a CRAN check and fixed some of the cube documentation.