Numerical Accuracy Update
Building the reproduction hazards had a numerical comparison issue. This
has an extra check now, with a tunable parameter propagated up to the
users.
This was not changed in the epi version, because there is a not about
our counting being off.
All of this stems from building an example script of SEM, which is still
in progress.