Update Default Parameter Handling
When running a network, it makes sense to internally handle parameter
expansion so that all nodes can be initialized with equal parameters and
be provided only a single number, instead of a vector of repeated
values. This has been handled for some parameters, but not all, as some
of the parameters are vectors/matrices and I don't think I should spend
that time on it right now.

The rest of the tweaks are me generating data to build analysis
functions. The V2 script doesn't work on a network - the releases are
wrong.