Many Updates
Fixed a release issue in the trajectory functions that was causing an
issue with the small-molecule style releases.
Also updated the documentation to be more useful.
Added SEM releases to the V1 test script.
Change the V2 test script to do allele counting over the population, as
an example to contrast the genotype counting of the V1 script.
I thought there was a problem with the cubes, there wasn't, but I did
fix some documentation.