Many Updates
I realized that there was a ton of code duplication in the decoupled
stuff. I added switches to some of the original functions to handle all
of the decoupled functionality, with proper defaults, and removed most
of the duplication. This applies to the node-P, node-T, hazard-main,
analysis, and plotting files or functions. Some of the original functions
were modified internally to make this switch easier.

The new equilibrium functions also had lots of duplication, which was
partly handled by changing original, internal functions for expanded
functionality. However, most of this duplication was left, as it isn't
my job currently and it won't be a simple, easy fix.

Vignettes have been updated per API changes from removing "decoupled"
functions from the namespace.

Lots of documentation was updated, but it is not completely correct for
the new functions. Again, going to take more time/effort than I can put
in at the moment.